# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Ugma Today API Documentation",
        "description": "",
        "license": {
            "name": "GNU GPL v3.0",
            "url": "https://www.gnu.org/licenses/gpl-3.0.en.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost/api",
            "description": "API Server"
        }
    ],
    "paths": {
        "/register": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Register new user",
                "description": "Create a user and return their data",
                "operationId": "register",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "username": {
                                        "description": "Username in the platform",
                                        "type": "string"
                                    },
                                    "password": {
                                        "description": "Password for login",
                                        "type": "string"
                                    },
                                    "name": {
                                        "description": "User's first name",
                                        "type": "string"
                                    },
                                    "lastname": {
                                        "description": "User's last name",
                                        "type": "string"
                                    },
                                    "email": {
                                        "description": "User's email",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "description": "Register new user",
                                            "properties": {
                                                "id": {
                                                    "description": "Register new user",
                                                    "type": "int"
                                                },
                                                "username": {
                                                    "description": "Register new user",
                                                    "type": "string"
                                                },
                                                "type": {
                                                    "description": "Register new user",
                                                    "type": "int"
                                                },
                                                "profile": {
                                                    "description": "Register new user",
                                                    "properties": {
                                                        "name": {
                                                            "description": "Register new user",
                                                            "type": "string"
                                                        },
                                                        "lastname": {
                                                            "description": "Register new user",
                                                            "type": "string"
                                                        },
                                                        "email": {
                                                            "description": "Register new user",
                                                            "type": "string"
                                                        },
                                                        "created_at": {
                                                            "description": "Register new user",
                                                            "type": "string"
                                                        },
                                                        "updated_at": {
                                                            "description": "Register new user",
                                                            "type": "string"
                                                        }
                                                    },
                                                    "type": "object"
                                                },
                                                "token": {
                                                    "description": "Register new user",
                                                    "type": "string"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Authenticate user",
                "description": "Authenticate user and return their data",
                "operationId": "login",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "username": {
                                        "description": "Authenticate user",
                                        "type": "string"
                                    },
                                    "password": {
                                        "description": "Authenticate user",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "description": "Authenticate user",
                                            "properties": {
                                                "id": {
                                                    "description": "Authenticate user",
                                                    "type": "int"
                                                },
                                                "username": {
                                                    "description": "Authenticate user",
                                                    "type": "string"
                                                },
                                                "type": {
                                                    "description": "Authenticate user",
                                                    "type": "int"
                                                },
                                                "profile": {
                                                    "description": "Authenticate user",
                                                    "properties": {
                                                        "name": {
                                                            "description": "Authenticate user",
                                                            "type": "string"
                                                        },
                                                        "lastname": {
                                                            "description": "Authenticate user",
                                                            "type": "string"
                                                        },
                                                        "email": {
                                                            "description": "Authenticate user",
                                                            "type": "string"
                                                        },
                                                        "created_at": {
                                                            "description": "Authenticate user",
                                                            "type": "string"
                                                        },
                                                        "updated_at": {
                                                            "description": "Authenticate user",
                                                            "type": "string"
                                                        }
                                                    },
                                                    "type": "object"
                                                },
                                                "token": {
                                                    "description": "Authenticate user",
                                                    "type": "string"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "Profile": {
                "title": "Profile model",
                "description": "Profile model",
                "properties": {
                    "name": {
                        "type": "string"
                    },
                    "lastname": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "User": {
                "title": "User model",
                "description": "User model",
                "properties": {
                    "username": {
                        "type": "string"
                    },
                    "type": {
                        "type": "int"
                    }
                },
                "type": "object"
            }
        }
    }
}